line renderer jitter
I changed the line renderer to local space calculations. it beams out from the gun tip and is always zero relative to the gun tip. the bit stuck to the wall probably jitters a little bit but its unnoticeable because I'm to busy swinging around to every notice it jitter small amounts. i used guntip.transform.inverseTransformPoint(gapplepoint.position);  to convert the unrelated world co-ordinates of the grapple point into local space for the gun tip