architecture, UI and animation
player is animated when dead. this needs to be adjusted. the code is cleaner and has better scalable architecture . a hitmarker now appears when shooting across the network. each player online has takes ownership of the ui local instance. changed transform instances to object intances and put respawn in the player .dead() method